+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+## Feature
+
+- **waffle-menu**: support for a configurable default icon to display (#59).
+- **waffle-meu**: support configurable truncation of text (#59).
+
 ## Fix
 
 - **portlet-registry-to-array**: ensure portlet array has unique `fname` (#57).
+- **content-carousel**: support `slick-options` in web component mode (#60).
 
 ## Test